callouts/requirements.txt
aiohttp==3.6.3
aioredis==1.0.0
aioscheduler==1.4.2
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
discord-ext-ipc==2.0.0
discord.py==1.5.1
emoji==0.6.0
hiredis==1.1.0
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
Pillow==8.0.1
psutil==5.6.6
pydest==0.4.0
pylint==2.6.0
PyMySQL==0.7.11
pytz==2017.2
redis==3.5.3
requests==2.24.0
schedule==0.6.0
selenium==3.141.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
wrapt==1.12.1
yarl==1.5.1
